Generic procedure for Neural HMM

+++ src/markov/discrete_time/controlled/abstract.jl
@@ -13,6 +13,8 @@ function transition_matrix(mc::AbstractControlledDiscreteMarkovChain, u, ps, st)
     return error("not implemented")
 end
 
-function transition_probability(mc::AbstractControlledDiscreteMarkovChain, i, j, u, ps, st)
+function transition_probability(
+    mc::AbstractControlledDiscreteMarkovChain, i::Integer, j::Integer, u, ps, st
+)
     return transition_matrix(mc, u, ps, st)[i, j]
 end
